# docs/deliverables/system_docs_20260729/star-rail-agents.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Star Rail Agents",
    "description": "基于 Multi-Agent 与 RAG 的游戏智能助手平台",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/planning/progression/calculate": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "合并计算多名已拥有角色的等级、技能和材料需求",
        "operationId": "calculate_progression_plan_api_v1_planning_progression_calculate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MultiProgressionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultiProgressionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/progression-plans": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "List Progression Plans",
        "operationId": "list_progression_plans_api_v1_profile_progression_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressionPlanListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "Create Progression Plan",
        "operationId": "create_progression_plan_api_v1_profile_progression_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressionPlanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressionPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/progression-plans/{plan_id}": {
      "patch": {
        "tags": [
          "planning"
        ],
        "summary": "Update Progression Plan",
        "operationId": "update_progression_plan_api_v1_profile_progression_plans__plan_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressionPlanUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressionPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "planning"
        ],
        "summary": "Delete Progression Plan",
        "operationId": "delete_progression_plan_api_v1_profile_progression_plans__plan_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planning/weekly/generate": {
      "post": {
        "tags": [
          "planning"
        ],
        "summary": "依据启用中的养成方案生成本周副本次数与体力分配",
        "operationId": "generate_weekly_plan_api_v1_planning_weekly_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyPlanGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/weekly-plans/current": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "Get Current Weekly Plan",
        "operationId": "get_current_weekly_plan_api_v1_profile_weekly_plans_current_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/profile/weekly-plans/{plan_id}/tasks/{task_id}": {
      "patch": {
        "tags": [
          "planning"
        ],
        "summary": "Update Weekly Task",
        "operationId": "update_weekly_task_api_v1_profile_weekly_plans__plan_id__tasks__task_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyTaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planning/daily": {
      "get": {
        "tags": [
          "planning"
        ],
        "summary": "查看滚动窗口内指定日期的每日体力安排（240 点上限，按优先级排程，周本置顶）",
        "operationId": "get_daily_plan_api_v1_planning_daily_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/planning/daily/{day}/{weekly_task_id}": {
      "patch": {
        "tags": [
          "planning"
        ],
        "summary": "勾选或取消某个日任务的完成状态，从当日 240 点账本核销并同步周计划进度",
        "operationId": "update_daily_task_api_v1_planning_daily__day___weekly_task_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Day"
            }
          },
          {
            "name": "weekly_task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Weekly Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyTaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MultiProgressionRequest": {
        "properties": {
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterPlanInput"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Characters"
          }
        },
        "type": "object",
        "required": [
          "characters"
        ],
        "title": "MultiProgressionRequest"
      },
      "MultiProgressionResponse": {
        "properties": {
          "response": {
            "$ref": "#/components/schemas/AIResponse"
          },
          "characters": {
            "items": {
              "$ref": "#/components/schemas/ProgressionCalculateResponse"
            },
            "type": "array",
            "title": "Characters"
          },
          "total_by_key": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Total By Key"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/ProgressionMaterialResponse"
            },
            "type": "array",
            "title": "Materials"
          },
          "recommendations": {
            "additionalProperties": {
              "$ref": "#/components/schemas/BuildRecommendation"
            },
            "type": "object",
            "title": "Recommendations"
          }
        },
        "type": "object",
        "required": [
          "response",
          "characters",
          "total_by_key",
          "materials",
          "recommendations"
        ],
        "title": "MultiProgressionResponse"
      },
...
      "ProgressionPlanCreate": {
        "properties": {
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterPlanInput"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Characters"
          },
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "priority": {
            "type": "integer",
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Priority",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "characters",
          "name"
        ],
        "title": "ProgressionPlanCreate"
      },
      "ProgressionPlanListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProgressionPlanResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ProgressionPlanListResponse"
      },
...
      "ProgressionPlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "completed"
            ],
            "title": "Status"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "request_payload": {
            "type": "object",
            "title": "Request Payload"
          },
          "material_snapshot": {
            "type": "object",
            "title": "Material Snapshot"
          },
          "recommendation_snapshot": {
            "type": "object",
            "title": "Recommendation Snapshot"
          },
          "progress": {
            "$ref": "#/components/schemas/ProgressionPlanProgress"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "priority",
          "request_payload",
          "material_snapshot",
          "recommendation_snapshot",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "ProgressionPlanResponse"
      },
      "ProgressionPlanUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 99.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "active",
                  "paused",
                  "completed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "ProgressionPlanUpdate"
      },
...
      "WeeklyPlanGenerateRequest": {
        "properties": {
          "stamina_budget": {
            "type": "integer",
            "maximum": 2400.0,
            "minimum": 0.0,
            "title": "Stamina Budget",
            "default": 1680
          },
          "weekly_runs_remaining": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Weekly Runs Remaining",
            "default": 3
          }
        },
        "type": "object",
        "title": "WeeklyPlanGenerateRequest"
      },
      "WeeklyPlanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "week_start": {
            "type": "string",
            "format": "date",
            "title": "Week Start"
          },
          "week_end": {
            "type": "string",
            "format": "date",
            "title": "Week End"
          },
          "stamina_budget": {
            "type": "integer",
            "title": "Stamina Budget"
          },
          "allocated_stamina": {
            "type": "integer",
            "title": "Allocated Stamina"
          },
          "weekly_runs_remaining": {
            "type": "integer",
            "title": "Weekly Runs Remaining"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/WeeklyTask"
            },
            "type": "array",
            "title": "Tasks"
          },
          "statistics": {
            "$ref": "#/components/schemas/WeeklyStatistics"
          },
          "notice": {
            "type": "string",
            "title": "Notice"
          },
          "evidence_version": {
            "type": "string",
            "title": "Evidence Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "week_start",
          "week_end",
          "stamina_budget",
          "allocated_stamina",
          "weekly_runs_remaining",
          "tasks",
          "statistics",
          "notice",
          "evidence_version",
          "created_at",
          "updated_at"
        ],
        "title": "WeeklyPlanResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地 FastAPI"
    }
  ]
}